# docs/swagger.yaml
definitions:
  model.Course:
    properties:
      categoryId:
        type: string
      course:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      uri:
        type: string
    type: object
  model.CourseCategory:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  model.Faculty:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  model.Session:
    properties:
      id:
        type: string
      userId:
        type: string
    type: object
...
info:
  contact:
    name: Jackie Li
  description: Syllabye API server.
  title: Syllabye API
  version: "1.0"
paths:
  /courses:
    get:
      parameters:
      - description: Search by course name or code
        in: query
        name: search
        type: string
      - description: Filter by category ID
        in: query
        name: category
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 10)'
        in: query
        name: size
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Course'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: List courses
      tags:
      - Course
  /courses/{courseId}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Course'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Get a course
      tags:
      - Course
  /courses/categories:
    get:
      parameters:
      - description: Search keyword
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.CourseCategory'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: List course categories
      tags:
      - Course Category
  /courses/categories/{categoryId}:
    get:
      parameters:
      - description: Category ID
        in: path
        name: categoryId
        required: true
        type: string
      responses:
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Retrieves a course category.
      tags:
      - Course Category
  /faculties:
    get:
      parameters:
      - description: Search by faculty name
        in: query
        name: search
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Faculty'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: List faculties
      tags:
      - Faculty
  /faculties/{facultyId}:
    get:
      parameters:
      - description: Faculty ID
        in: path
        name: facultyId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Faculty'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - Session: []
      summary: Get a faculty
      tags:
      - Faculty
  /me:
    get:
      description: Validates the session cookie and returns session payload if authenticated.
      responses:
        "200":
          description: Valid session
          schema:
            $ref: '#/definitions/model.Session'
        "401":
          description: Missing or invalid session cookie
          schema:
            type: string
      security:
      - Session: []
      summary: Check user session
      tags:
      - Authentication
...
  /providers/google:
    get:
      description: Validates an optional redirect query param and redirects the user
        to the OpenID login flow.
      parameters:
      - description: Optional redirect URL after login
        in: query
        name: redirect
        type: string
      responses:
        "302":
          description: Redirects to OpenID consent screen
          schema:
            type: string
        "500":
          description: Unable to continue to OpenID provider
          schema:
            type: string
      summary: Redirect to OpenID consent screen
      tags:
      - Authentication
  /providers/google/callback:
    get:
      description: Validates authorization code and state, registers or logs in user,
        and sets session cookie.
      parameters:
      - description: Authorization code returned by the OpenID provider
        in: query
        name: code
        required: true
        type: string
      - description: State token for CSRF protection and redirect tracking
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Redirects to dashboard or original destination
          schema:
            type: string
        "401":
          description: Invalid or expired state token
          schema:
            type: string
        "500":
          description: Token exchange, validation, or session creation failed
          schema:
            type: string
      summary: OpenID provider callback
      tags:
      - Authentication
...
securityDefinitions:
  AWS:
    description: 'Provide a valid Bearer token. Example: "Bearer {jwt-token}"'
    in: header
    name: Authorization
    type: apiKey
  Session:
    in: cookie
    name: syllabye.session
    type: apiKey
swagger: "2.0"